remove create post component
